+++ lib/utils/consts.dart
@@ -1,0 +1,1 @@
+const showBrightSchemeFlagName = 'show_bright_scheme';