+++ lib/widgets/misc/tasks_counts.dart
@@ -69,8 +69,8 @@ class TasksCountWidget extends StatelessWidget {
             padding: const EdgeInsets.symmetric(horizontal: 4),
             child: Text(
               '$label: ${snapshot.data}',
-              style: TextStyle(
-                color: AppColors.headerFontColor2,
+              style: const TextStyle(
+                color: AppColors.headerFontColor,
                 fontFamily: 'Oswald',
                 fontSize: 12,
                 fontWeight: FontWeight.w100,