+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added:
+- Bright ☀️ color scheme with config flag
+- Change️ color scheme from menu on desktop
+
+## [0.8.80] - 2022-06-27
 ### Changed:
 - Use AppRouter mock in tests (no UI changes)